Changed i.q parsing

+++ sheet2json/sheet2json.py
@@ -44,7 +44,7 @@ def process_i_p(input):
 
 def process_i_q(input):
     """Return quantity (i.q) from input string."""
-    return int(input)  # TODO
+    return float(input)  # TODO
 
 
 def process_i_s(input):
